test(native-accordion): recategorize and clean up browser tests
